fs:change uiomove() to use size_t in the signature instead of int
As Wonsup Yoon reported in the issue #1095 following C++ trying
to read large number of bytes from /dev/random fails:

```
#include<fstream>
#include<vector>
#include<iostream>

#define READ_BYTES (1LL << 31)

int main(){

	std::ifstream file("/dev/urandom");

	std::vector<char> data(READ_BYTES);
	data[0] = 1;
	file.read(data.data(), READ_BYTES);

	std::cout << +data[0] << std::endl;

	return 0;
}
```

To make long story short (see the issue for more details)
it turns out that the culprit is wron signature of uiomove()
function that uses int instead of size_t for n parameter.
This patch fixes it.

Fixes #1095

+++ fs/vfs/subr_uio.cc
@@ -41,13 +41,13 @@
 #include <osv/uio.h>
 
 int
-uiomove(void *cp, int n, struct uio *uio)
+uiomove(void *cp, size_t n, struct uio *uio)
 {
 	assert(uio->uio_rw == UIO_READ || uio->uio_rw == UIO_WRITE);
 
 	while (n > 0 && uio->uio_resid) {
 		struct iovec *iov = uio->uio_iov;
-		int cnt = iov->iov_len;
+		size_t cnt = iov->iov_len;
 		if (cnt == 0) {
 			uio->uio_iov++;
 			uio->uio_iovcnt--;
